Specify dtype for bond indices

+++ timemachine/md/barostat/utils.py
@@ -41,7 +41,7 @@ def get_group_indices(bond_list: list[tuple[int, int]], num_atoms: int) -> list[
     """Connected components of bond graph"""
 
     topology = nx.Graph(bond_list)
-    components = [np.array(list(sorted(c))) for c in nx.algorithms.connected_components(topology)]
+    components = [np.array(list(sorted(c)), dtype=np.int32) for c in nx.algorithms.connected_components(topology)]
     for i in range(len(components)):
         assert np.all(np.diff(components[i]) == 1)
